+++ src/Curl/RequestMethodEnum.php
@@ -10,27 +10,57 @@ class RequestMethodEnum extends Enum
     public const POST = 'POST';
     public const PUT = 'PUT';
     public const DELETE = 'DELETE';
+    public const PATCH = 'PATCH';
+    public const OPTIONS = 'OPTIONS';
 
     //--------------------------------------------------------------------------
     // These methods are just for IDE autocomplete and not are mandatory.
     //--------------------------------------------------------------------------
+    /**
+     * @return static
+     */
     public static function GET(): self
     {
-        return new self (self::GET);
+        return new static (static::GET);
     }
 
+    /**
+     * @return static
+     */
     public static function POST(): self
     {
-        return new self (self::POST);
+        return new static (static::POST);
     }
 
+    /**
+     * @return static
+     */
     public static function PUT(): self
     {
-        return new self (self::PUT);
+        return new static (static::PUT);
     }
 
+    /**
+     * @return static
+     */
     public static function DELETE(): self
     {
-        return new self (self::DELETE);
+        return new static (static::DELETE);
+    }
+
+    /**
+     * @return static
+     */
+    public static function PATCH(): self
+    {
+        return new static (static::PATCH);
+    }
+
+    /**
+     * @return static
+     */
+    public static function OPTIONS(): self
+    {
+        return new static (static::OPTIONS);
     }
 }